backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
numpy
orjson
sse-starlette
python-multipart
# transformers
FlagEmbedding
torch